fix(signal): align struct sigaction to Linux kernel ABI, add SA_SIGINFO
Reorder struct sigaction fields to match Linux rt_sigaction kernel ABI:
  [+0]  sa_handler / sa_sigaction (union)
  [+8]  sa_flags (unsigned long, was int after sa_mask)
  [+16] sa_restorer (new field for libc signal trampoline)
  [+24] sa_mask (sigset_t, was at +8)

Also:
- Add SA_SIGINFO (0x4), SA_NOCLDWAIT (0x2), SA_RESTORER (0x04000000)
- Add sa_sigaction union member for 3-arg SA_SIGINFO handlers
- Move sigset_t definition before struct sigaction (fixes forward-ref)
- Change signal_handler_flags[] from int[] to unsigned long[] to match
  the unsigned long sa_flags type in the Linux ABI

Without this fix, glibc/musl rt_sigaction calls would have sa_flags and
sa_mask swapped, causing signal handlers to run with wrong masks/flags.

+++ kernel/signal/signal.c
@@ -372,7 +372,7 @@ int fut_signal_deliver(struct fut_task *task, void *frame) {
 
     /* Phase 2: Select signal stack — use alternate stack if SA_ONSTACK set
      * and the alternate stack is active (not disabled, not already on it). */
-    int handler_flags = task->signal_handler_flags[signum - 1];
+    unsigned long handler_flags = task->signal_handler_flags[signum - 1];
     uint64_t sp;
     if ((handler_flags & SA_ONSTACK) &&
         !(task->sig_altstack.ss_flags & SS_DISABLE) &&
@@ -667,7 +667,7 @@ int fut_signal_deliver(struct fut_task *task, void *frame) {
 
     /* Block signals during handler: sa_mask + the delivered signal itself.
      * SA_NODEFER: don't block the delivered signal (allow recursive delivery). */
-    int handler_flags_x86 = task->signal_handler_flags[signum - 1];
+    unsigned long handler_flags_x86 = task->signal_handler_flags[signum - 1];
     uint64_t new_mask_bits_x86 = task->signal_handler_masks[signum - 1] &
                                  ~((1ULL << (SIGKILL - 1)) | (1ULL << (SIGSTOP - 1)));
     if (!(handler_flags_x86 & SA_NODEFER)) {
